+++ src/borrower/feed/test/navfeedPV.t.sol
@@ -7,55 +7,57 @@ import "./../../../../test/borrower/mock/shelf.sol";
 import "./../../../../test/borrower/mock/pile.sol";
 import "../navfeedPV.sol";
 
-
 contract NFTFeedTest is DSTest, Math {
     NAVFeedPV public feed;
     ShelfMock shelf;
     PileMock pile;
 
-    uint defaultRate;
-    uint defaultThresholdRatio;
-    uint defaultCeilingRatio;
+    uint256 defaultRate;
+    uint256 defaultThresholdRatio;
+    uint256 defaultCeilingRatio;
 
     function setUp() public {
         // default values
-        defaultThresholdRatio = 8*10**26;                     // 80% threshold
-        defaultCeilingRatio = 6*10**26;                       // 60% ceiling
-        defaultRate = uint(1000000564701133626865910626);     // 5 % day
+        defaultThresholdRatio = 8 * 10 ** 26; // 80% threshold
+        defaultCeilingRatio = 6 * 10 ** 26; // 60% ceiling
+        defaultRate = uint256(1000000564701133626865910626); // 5 % day
 
         feed = new NAVFeedPV();
         pile = new PileMock();
         shelf = new ShelfMock();
         feed.depend("shelf", address(shelf));
         feed.depend("pile", address(pile));
-        
-        feed.file("riskGroup",
-            0,                                      // riskGroup:       0
-            8*10**26,                               // thresholdRatio   70%
-            6*10**26,                               // ceilingRatio     60%
-            uint(1000000564701133626865910626)      // interestRate     5% per year
+
+        feed.file(
+            "riskGroup",
+            0, // riskGroup:       0
+            8 * 10 ** 26, // thresholdRatio   70%
+            6 * 10 ** 26, // ceilingRatio     60%
+            uint256(1000000564701133626865910626) // interestRate     5% per year
         );
 
-         feed.file("riskGroup",
-            1,                                      // riskGroup:       0
-            8*10**26,                               // thresholdRatio   80%
-            7*10**26,                               // ceilingRatio     70%
-            ONE                                     // interestRate     1.0
+        feed.file(
+            "riskGroup",
+            1, // riskGroup:       0
+            8 * 10 ** 26, // thresholdRatio   80%
+            7 * 10 ** 26, // ceilingRatio     70%
+            ONE // interestRate     1.0
         );
-        feed.file("riskGroup",
-            2,                                      // riskGroup:       0
-            8*10**26,                               // thresholdRatio   90%
-            8*10**26,                               // ceilingRatio     80%
-            uint(1000000003593629043335673583)      // interestRate     12% per year
+        feed.file(
+            "riskGroup",
+            2, // riskGroup:       0
+            8 * 10 ** 26, // thresholdRatio   90%
+            8 * 10 ** 26, // ceilingRatio     80%
+            uint256(1000000003593629043335673583) // interestRate     12% per year
         );
     }
 
     function testNFTValues() public {
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint value = 100 ether;
+        uint256 value = 100 ether;
         feed.update(nftID, value);
 
-        uint loan = 1;
+        uint256 loan = 1;
         shelf.setReturn("shelf", address(1), 1);
 
         assertEq(feed.nftValues(nftID), 100 ether);
@@ -65,11 +67,11 @@ contract NFTFeedTest is DSTest, Math {
 
     function testCreditLineCeilingOutstandingDebt() public {
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint riskGroup = 1;
-        uint value = 100 ether;
-        uint loan = 1;
-        uint debt = 40 ether;
-        uint maxCeiling = 70 ether;
+        uint256 riskGroup = 1;
+        uint256 value = 100 ether;
+        uint256 loan = 1;
+        uint256 debt = 40 ether;
+        uint256 maxCeiling = 70 ether;
         feed.update(nftID, value, riskGroup);
 
         shelf.setReturn("shelf", address(1), 1);
@@ -80,11 +82,11 @@ contract NFTFeedTest is DSTest, Math {
 
     function testCreditLineCeilingOutstandingDebtBiggerCeiling() public {
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint riskGroup = 1;
-        uint loan = 1;
-        uint value = 100 ether;
-        uint debt = 71 ether;
-        uint maxCeiling = 70 ether;
+        uint256 riskGroup = 1;
+        uint256 loan = 1;
+        uint256 value = 100 ether;
+        uint256 debt = 71 ether;
+        uint256 maxCeiling = 70 ether;
         feed.update(nftID, value, riskGroup);
 
         shelf.setReturn("shelf", address(1), 1);
@@ -94,42 +96,38 @@ contract NFTFeedTest is DSTest, Math {
     }
 
     function testAddRiskGroup() public {
-        uint riskGroup = 4;
-        uint thresholdRatio = 8*10**26;
-        uint ceilingRatio =  8*10**26; 
-        uint interestRate =  uint(1000000003593629043335673583);
+        uint256 riskGroup = 4;
+        uint256 thresholdRatio = 8 * 10 ** 26;
+        uint256 ceilingRatio = 8 * 10 ** 26;
+        uint256 interestRate = uint256(1000000003593629043335673583);
 
-        feed.file("riskGroup",
-            riskGroup,                                  
-            thresholdRatio,                              
-            ceilingRatio,
-            interestRate                              
-        );
-        (uint pie, uint chi, uint ratePerSecond, uint lastUpdate, uint fixedRate) = pile.rates(riskGroup);
+        feed.file("riskGroup", riskGroup, thresholdRatio, ceilingRatio, interestRate);
+        (uint256 pie, uint256 chi, uint256 ratePerSecond, uint256 lastUpdate, uint256 fixedRate) = pile.rates(riskGroup);
 
-         // check nft value set correctly
+        // check nft value set correctly
         assertEq(feed.thresholdRatio(riskGroup), thresholdRatio);
         // check threshold computed correctly
         assertEq(feed.ceilingRatio(riskGroup), ceilingRatio);
-         // check interestrate set correctly
+        // check interestrate set correctly
         assertEq(ratePerSecond, interestRate);
     }
 
     function testFailAddRiskGroupNoPermissions() public {
         feed.deny(address(this)); // revoke admin permissions
-        feed.file("riskGroup",
-            2,                                      // riskGroup:       0
-            8*10**26,                               // thresholdRatio   90%
-            8*10**26,                               // ceilingRatio     80%
-            uint(1000000003593629043335673583)      // interestRate     12% per year
+        feed.file(
+            "riskGroup",
+            2, // riskGroup:       0
+            8 * 10 ** 26, // thresholdRatio   90%
+            8 * 10 ** 26, // ceilingRatio     80%
+            uint256(1000000003593629043335673583) // interestRate     12% per year
         );
     }
 
     function testUpdateRiskGroupLoanHasNoDebt() public {
-        uint risk = 1;
+        uint256 risk = 1;
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint loan = 1;
-        uint value = 100 ether;
+        uint256 loan = 1;
+        uint256 value = 100 ether;
         shelf.setReturn("shelf", address(1), 1);
         // set value and risk group of nft
         feed.update(nftID, value, risk);
@@ -137,17 +135,17 @@ contract NFTFeedTest is DSTest, Math {
         assertLoanValuesSetCorrectly(nftID, value, loan, risk, false);
 
         // set new riskGroup
-        uint defaultRisk = 0;
+        uint256 defaultRisk = 0;
         feed.update(nftID, value, defaultRisk);
         // assert all values were set correctly
         assertLoanValuesSetCorrectly(nftID, value, loan, defaultRisk, false);
     }
 
     function testUpdateRiskGroupAndValueLoanHasDebt() public {
-        uint risk = 1;
+        uint256 risk = 1;
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint loan = 1;
-        uint value = 100 ether;
+        uint256 loan = 1;
+        uint256 value = 100 ether;
         shelf.setReturn("shelf", address(1), 1);
         shelf.setReturn("nftlookup", loan);
         pile.setReturn("pie", 100);
@@ -168,10 +166,10 @@ contract NFTFeedTest is DSTest, Math {
     function testFailUpdateRiskGroupNoPermissions() public {
         feed.deny(address(this)); // revoke admin permissions
 
-        uint risk = 1;
+        uint256 risk = 1;
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint loan = 1;
-        uint value = 100 ether;
+        uint256 loan = 1;
+        uint256 value = 100 ether;
         shelf.setReturn("shelf", address(1), 1);
         shelf.setReturn("nftlookup", loan);
         pile.setReturn("pie", 100);
@@ -182,10 +180,10 @@ contract NFTFeedTest is DSTest, Math {
 
     function testBorrowEvent() public {
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint value = 100 ether;
-        uint risk = 0;
+        uint256 value = 100 ether;
+        uint256 risk = 0;
 
-        uint loan = 1;
+        uint256 loan = 1;
         shelf.setReturn("shelf", address(1), 1);
         pile.setReturn("loanRates", 1);
 
@@ -199,11 +197,11 @@ contract NFTFeedTest is DSTest, Math {
 
     function testBorrow() public {
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint riskGroup = 1;
-        uint value = 100 ether;
-        uint loan = 1;
-        uint debt = 40 ether;
-        uint maxCeiling = 70 ether;
+        uint256 riskGroup = 1;
+        uint256 value = 100 ether;
+        uint256 loan = 1;
+        uint256 debt = 40 ether;
+        uint256 maxCeiling = 70 ether;
 
         feed.update(nftID, value, riskGroup);
 
@@ -212,14 +210,14 @@ contract NFTFeedTest is DSTest, Math {
 
         feed.borrow(loan, maxCeiling);
     }
-    
+
     function testFailBorrowAmountTooHigh() public {
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint riskGroup = 1;
-        uint value = 100 ether;
-        uint loan = 1;
-        uint debt = 40 ether;
-        uint maxCeiling = 70 ether;
+        uint256 riskGroup = 1;
+        uint256 value = 100 ether;
+        uint256 loan = 1;
+        uint256 debt = 40 ether;
+        uint256 maxCeiling = 70 ether;
 
         feed.update(nftID, value, riskGroup);
 
@@ -232,11 +230,11 @@ contract NFTFeedTest is DSTest, Math {
 
     function testFailBorrowNoPermission() public {
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint riskGroup = 1;
-        uint value = 100 ether;
-        uint loan = 1;
-        uint debt = 40 ether;
-        uint maxCeiling = 70 ether;
+        uint256 riskGroup = 1;
+        uint256 value = 100 ether;
+        uint256 loan = 1;
+        uint256 debt = 40 ether;
+        uint256 maxCeiling = 70 ether;
 
         feed.update(nftID, value, riskGroup);
 
@@ -247,15 +245,15 @@ contract NFTFeedTest is DSTest, Math {
         feed.borrow(loan, maxCeiling);
     }
 
-     function testWriteOff() public {
-        uint risk = 1;
+    function testWriteOff() public {
+        uint256 risk = 1;
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint loan = 1;
-        uint value = 100 ether;
+        uint256 loan = 1;
+        uint256 value = 100 ether;
         shelf.setReturn("shelf", address(1), 1);
         shelf.setReturn("nftlookup", loan);
         pile.setReturn("pie", 100);
-    
+
         // set value and risk group of nft
         feed.update(nftID, value, risk);
         // assert all values were set correctly
@@ -267,14 +265,14 @@ contract NFTFeedTest is DSTest, Math {
     }
 
     function testFailWriteOffNoPermissions() public {
-        uint risk = 1;
+        uint256 risk = 1;
         bytes32 nftID = feed.nftID(address(1), 1);
-        uint loan = 1;
-        uint value = 100 ether;
+        uint256 loan = 1;
+        uint256 value = 100 ether;
         shelf.setReturn("shelf", address(1), 1);
         shelf.setReturn("nftlookup", loan);
         pile.setReturn("pie", 100);
-    
+
         // set value and risk group of nft
         feed.update(nftID, value, risk);
         // assert all values were set correctly
@@ -288,41 +286,47 @@ contract NFTFeedTest is DSTest, Math {
         bytes32 tokenId1 = feed.nftID(address(1), 1);
         bytes32 tokenId2 = feed.nftID(address(1), 2);
         bytes32 tokenId3 = feed.nftID(address(1), 3);
-        uint value1 = 100;
-        uint value2 = 1000;
-        uint value3 = 10000;
-        uint loan1 = 1;
-        uint loan2 = 2;
-        uint loan3 = 3;
-        uint riskGroup = 1;
-    
+        uint256 value1 = 100;
+        uint256 value2 = 1000;
+        uint256 value3 = 10000;
+        uint256 loan1 = 1;
+        uint256 loan2 = 2;
+        uint256 loan3 = 3;
+        uint256 riskGroup = 1;
+
         feed.update(tokenId1, value1, riskGroup);
         feed.update(tokenId2, value2, riskGroup);
         feed.update(tokenId3, value3, riskGroup);
 
         shelf.setReturn("shelf", address(1), 1);
         feed.borrow(loan1, feed.ceiling(loan1));
-        pile.incDebt(loan1, feed.ceiling(loan1)); 
+        pile.incDebt(loan1, feed.ceiling(loan1));
         shelf.setReturn("shelf", address(1), 2);
         feed.borrow(loan2, feed.ceiling(loan2));
-        pile.incDebt(loan2,feed.ceiling(loan2)); 
+        pile.incDebt(loan2, feed.ceiling(loan2));
         shelf.setReturn("shelf", address(1), 3);
         feed.borrow(loan3, feed.ceiling(loan3));
-        pile.incDebt(loan3, feed.ceiling(loan3)); 
+        pile.incDebt(loan3, feed.ceiling(loan3));
 
         shelf.setReturn("loanCount", 4);
-        uint currentNav = feed.currentNAV();
+        uint256 currentNav = feed.currentNAV();
         assertEq(currentNav, safeAdd(safeAdd(70, 700), 7000));
     }
 
     // assertions
-    function assertLoanValuesSetCorrectly(bytes32 nftID, uint nftValue, uint loan, uint riskGroup, bool loanHasDebt) public {
+    function assertLoanValuesSetCorrectly(
+        bytes32 nftID,
+        uint256 nftValue,
+        uint256 loan,
+        uint256 riskGroup,
+        bool loanHasDebt
+    ) public {
         // check nft value set correctly
         assertEq(feed.nftValues(nftID), nftValue);
         // check threshold computed correctly
         assertEq(feed.threshold(loan), rmul(nftValue, feed.thresholdRatio(riskGroup)));
         // check ceiling computed correctly
-        assertEq(feed.ceiling(loan),  rmul(nftValue, feed.ceilingRatio(riskGroup)));
+        assertEq(feed.ceiling(loan), rmul(nftValue, feed.ceilingRatio(riskGroup)));
         // check rate set correctly in pile
 
         if (loanHasDebt) {
@@ -331,5 +335,3 @@ contract NFTFeedTest is DSTest, Math {
         }
     }
 }
-
-
